fix profile recheck error
