+++ test/e2e/e2e-service-provider/src/main/java/test/apache/skywalking/e2e/Service0Application.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.skywalking.e2e;
+package test.apache.skywalking.e2e;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
