+++ test/e2e/e2e-test/src/test/java/org/apache/skywalking/e2e/profile/ProfileE2E.java
@@ -141,7 +141,7 @@ public class ProfileE2E extends SkyWalkingTestAdapter {
     void createProfileTask() throws Exception {
         final ProfileTaskCreationRequest creationRequest = ProfileTaskCreationRequest.builder()
                                                                                      .serviceId("ZTJlLXByb2ZpbGUtc2VydmljZQ==.1")
-                                                                                     .endpointName("/profile/users")
+                                                                                     .endpointName("/profile/{name}")
                                                                                      .duration(1)
                                                                                      .startTime(-1)
                                                                                      .minDurationThreshold(1500)
