+++ test/e2e/e2e-test/src/test/resources/expected/profile/notified.yml
@@ -16,7 +16,7 @@
 tasks:
   - id: not null
     serviceId: ZTJlLXByb2ZpbGUtc2VydmljZQ==.1
-    endpointName: /profile/users
+    endpointName: /profile/{name}
     startTime: gt 0
     duration: gt 0
     minDurationThreshold: gt 0
