+++ test/e2e/e2e-test/src/test/resources/expected/profile/profileSegment.yml
@@ -19,7 +19,7 @@ spans:
     serviceCode: not null
     startTime: gt 0
     endTime: gt 0
-    endpointName: /profile/users
+    endpointName: /profile/{name}
     tags:
       - key: url
         value: not null
